Change remaining IDs to uint64

+++ lib/components/include/circsim/components/Wire.hpp
@@ -128,10 +128,10 @@ private:
     extern_func_t _driver_function;
 
     /// The set of transistors which control this wire
-    std::vector<size_t> _trans_ctl_ids;
+    std::vector<uint64_t> _trans_ctl_ids;
 
     /// The set of transistors controlled in turn by this wire
-    std::vector<size_t> _trans_gate_ids;
+    std::vector<uint64_t> _trans_gate_ids;
 
 public:
 
@@ -154,10 +154,10 @@ public:
      */
     Wire
     (
-        const size_t id,
+        const uint64_t id,
         const SpecialWireType special_type,
-        const std::vector<size_t> &control_transistors,
-        const std::vector<size_t> &gate_transistors
+        const std::vector<uint64_t> &control_transistors,
+        const std::vector<uint64_t> &gate_transistors
     );
 
     /**
@@ -171,11 +171,11 @@ public:
      */
     Wire
     (
-        const size_t id,
+        const uint64_t id,
         const std::string &name,
         const std::function<State(State)> driver_func,
-        const std::vector<size_t> &control_transistors,
-        const std::vector<size_t> &gate_transistors
+        const std::vector<uint64_t> &control_transistors,
+        const std::vector<uint64_t> &gate_transistors
     );
 
     /**
@@ -189,11 +189,11 @@ public:
      */
     Wire
     (
-        const size_t id,
+        const uint64_t id,
         const std::string &name,
         const PulledStatus pulled,
-        const std::vector<size_t> &control_transistors,
-        const std::vector<size_t> &gate_transistors
+        const std::vector<uint64_t> &control_transistors,
+        const std::vector<uint64_t> &gate_transistors
     );
 
 
@@ -275,16 +275,16 @@ public:
     /**
      * @brief Return the list of control transistors for this object.
      * 
-     * @return std::vector<size_t> The IDs of the control transistors.
+     * @return std::vector<uint64_t> The IDs of the control transistors.
      */
-    inline std::vector<size_t> ctrl_transistors() const noexcept { return _trans_ctl_ids; }
+    inline std::vector<uint64_t> ctrl_transistors() const noexcept { return _trans_ctl_ids; }
 
     /**
      * @brief Return the list of gate transistors for this wire.
      * 
-     * @return std::vector<size_t> The IDs of the gate transistor.
+     * @return std::vector<uint64_t> The IDs of the gate transistor.
      */
-    inline std::vector<size_t> gate_transistors() const noexcept { return _trans_gate_ids; }
+    inline std::vector<uint64_t> gate_transistors() const noexcept { return _trans_gate_ids; }
 
     /**
      * @brief Is this wire driven externally, or by other transistors in the circuit.
